+++ src/d-counter/DCounter.stories.js
@@ -37,6 +37,11 @@ const DarkModeTemplate = (args) => ({
 
 export const Default = Template.bind({});
 
+export const Label = Template.bind({});
+Label.args = {
+  label: "Counter Input",
+};
+
 export const HasError = Template.bind({});
 HasError.args = {
   errorMessage: "Error message",