+++ src/d-date-picker/DDatePicker.stories.js
@@ -84,6 +84,11 @@ const DarkModeTemplateFactory = () => {
 
 export const Default = Template.bind({});
 
+export const Label = Template.bind({});
+Label.args = {
+  label: "Date picker",
+};
+
 export const HasError = Template.bind({});
 HasError.args = {
   errorMessage: "Error message",